+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 *   (improvement) Bump required node version to 12.
 *   (internal) Bumped all dependencies.
 *   (improvement) Allow `++` in JS/TS code.
+*   (feature) Enable separate compiler instances for each entry file by using `enablePerEntryCompilation()`
 
 
 8.1.0